Fixed deleting all duties

+++ backend/src/main/java/org/agh/backend/service/DutyService.java
@@ -6,6 +6,7 @@ import org.agh.backend.model.Appointment;
 import org.agh.backend.model.Doctor;
 import org.agh.backend.model.Duty;
 import org.agh.backend.model.Office;
+import org.agh.backend.repository.AppointmentRepository;
 import org.agh.backend.repository.DoctorRepository;
 import org.agh.backend.repository.DutyRepository;
 import org.agh.backend.repository.OfficeRepository;
@@ -19,15 +20,18 @@ public class DutyService {
     private final DutyRepository dutyRepository;
     private final DoctorRepository doctorRepository;
     private final OfficeRepository officeRepository;
+    private final AppointmentRepository appointmentRepository;
 
     public DutyService(
             DutyRepository dutyRepository,
             DoctorRepository doctorRepository,
-            OfficeRepository officeRepository
+            OfficeRepository officeRepository,
+            AppointmentRepository appointmentRepository
     ) {
         this.dutyRepository = dutyRepository;
         this.doctorRepository = doctorRepository;
         this.officeRepository = officeRepository;
+        this.appointmentRepository = appointmentRepository;
     }
 
     /**
@@ -102,8 +106,12 @@ public class DutyService {
 
     /**
      * Deletes all duties
+     * @throws IllegalStateException if any duty has scheduled appointment
      */
     public void deleteAllDuties() {
+        if (!appointmentRepository.findAll().isEmpty()) {
+            throw new IllegalStateException("Some duties have scheduled appointments");
+        }
         dutyRepository.deleteAll();
     }
 }
